check the rule
check the first part whether 'READ', 'GET', 'PUT' .Check whether the two conditions are met: The minimum size is 7 (which is the case when the key in a READ/GET operation is a single character), and the maximum size is 999. As a rule, the "collated size" of the key and value (represented as a single string separated by spaces) should not exceed 970 characters.